remove process_EPR from tools; unittest for 'null'
